+++ src/main/scala/com/codahale/simplespec/Spec.scala
@@ -3,6 +3,7 @@ package com.codahale.simplespec
 import org.specs2.mutable.Specification
 import org.specs2.execute.Result
 import org.specs2.specification.Step
+import java.lang.reflect.InvocationTargetException
 
 trait Spec extends Specification with Discovery {
   def beforeAll() {}
@@ -17,9 +18,19 @@ trait Spec extends Specification with Discovery {
         requirements.map { req =>
           req.name in {
             req.evaluate() match {
+              case null => success
               case r: Result => r
-              case r: { def toResult: Result } => r.toResult
-              case _ => pending
+              case other: Object => {
+                try {
+                  val result = other.getClass.getMethod("toResult")
+                  if (classOf[Result].isAssignableFrom(result.getReturnType)) {
+                    result.invoke(other).asInstanceOf[Result]
+                  } else pending
+                } catch {
+                  case e@(_: NoSuchMethodException | _:InvocationTargetException) => pending
+                }
+              }
+              case unknown => pending
             }
           }
         }.head